feat: InputText 와 InputContainer 분리 및 폴더 구조 정리
