+++ client/src/components/Input/InputText/InputText.jsx
@@ -1,0 +1,20 @@
+import styles from './InputText.module.css';
+
+export default function InputText({
+  type = 'text',
+  name,
+  placeholder,
+  value,
+  onChange,
+}) {
+  return (
+    <input
+      className={styles.input}
+      name={name}
+      type={type}
+      placeholder={placeholder}
+      value={value}
+      onChange={onChange}
+    />
+  );
+}
