+++ client/src/pages/Login/Login.jsx
@@ -1,7 +1,7 @@
 import styles from './Login.module.css';
 import { instance } from '../../apis/instance.js';
 import { useState } from 'react';
-import InputText from '../../components/InputText/InputText';
+import InputText from '../../components/Input/InputText/InputText';
 import BtnPrimary from '../../components/BtnPrimary/BtnPrimary';
 
 export default function Login() {
